+++ docs/testdata/TestGeneratePluginDocs-JSON-__tables.json
