+++ plugin/options.go
@@ -1,8 +1,6 @@
 package plugin
 
 import (
-	"fmt"
-
 	"github.com/cloudquery/plugin-sdk/v4/schema"
 )
 
@@ -19,31 +17,6 @@ var (
 
 func (m MigrateMode) String() string {
 	return migrateModeStrings[m]
-}
-
-type Registry int
-
-const (
-	RegistryGithub Registry = iota
-	RegistryLocal
-	RegistryGrpc
-)
-
-func (r Registry) String() string {
-	return [...]string{"github", "local", "grpc"}[r]
-}
-
-func RegistryFromString(s string) (Registry, error) {
-	switch s {
-	case "github":
-		return RegistryGithub, nil
-	case "local":
-		return RegistryLocal, nil
-	case "grpc":
-		return RegistryGrpc, nil
-	default:
-		return RegistryGithub, fmt.Errorf("unknown registry %s", s)
-	}
 }
 
 type WriteMode int
@@ -63,7 +36,6 @@ func (m WriteMode) String() string {
 }
 
 type Option func(*Plugin)
-
 
 // WithNoInternalColumns won't add internal columns (_cq_id, _cq_parent_cq_id) to the plugin tables
 func WithNoInternalColumns() Option {
