+++ serve/plugin.go
@@ -245,36 +245,6 @@ func (s *PluginServe) newCmdPluginServe() *cobra.Command {
 	return cmd
 }
 
-const (
-	pluginDocShort = "Generate documentation for tables"
-	pluginDocLong  = `Generate documentation for tables
-
-If format is markdown, a destination directory will be created (if necessary) containing markdown files.
-Example:
-doc ./output 
-
-If format is JSON, a destination directory will be created (if necessary) with a single json file called __tables.json.
-Example:
-doc --format json .
-`
-)
-
-func (s *PluginServe) newCmdPluginDoc() *cobra.Command {
-	format := newEnum([]string{"json", "markdown"}, "markdown")
-	cmd := &cobra.Command{
-		Use:   "doc <directory>",
-		Short: pluginDocShort,
-		Long:  pluginDocLong,
-		Args:  cobra.ExactArgs(1),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			pbFormat := pbv3.GenDocs_FORMAT(pbv3.GenDocs_FORMAT_value[format.Value])
-			return s.plugin.GeneratePluginDocs(args[0], pbFormat)
-		},
-	}
-	cmd.Flags().Var(format, "format", fmt.Sprintf("output format. one of: %s", strings.Join(format.Allowed, ",")))
-	return cmd
-}
-
 func (s *PluginServe) newCmdPluginRoot() *cobra.Command {
 	cmd := &cobra.Command{
 		Use: fmt.Sprintf("%s <command>", s.plugin.Name()),
